remove qwen2.py
